Move structure to different files
