+++ src/TypeDescriptor.cpp
@@ -1,0 +1,10 @@
+#include "TypeDescriptor.hpp"
+
+// Define member functions for the TypeDescriptor class here, if needed
+TypeDescriptor::TypeDescriptor() {
+    // Constructor implementation
+}
+
+TypeDescriptor::~TypeDescriptor() {
+    // Destructor implementation
+}
