+++ src/TypeDescriptor.hpp
@@ -1,0 +1,15 @@
+#ifndef TYPEDESCRIPTOR_HPP
+#define TYPEDESCRIPTOR_HPP
+
+class TypeDescriptor {
+	public:
+		// Declare member functions, data members, or any other class-specific elements here
+
+		// For example:
+		TypeDescriptor();  // Constructor declaration
+		~TypeDescriptor(); // Destructor declaration
+
+		// Additional member function declarations, if needed
+};
+
+#endif // TYPEDESCRIPTOR_HPP
